99-11-2

+++ app/Http/Controllers/AdminUsersController.php
@@ -10,6 +10,8 @@ use App\Http\Requests\UserEditRequest;
 use App\Models\Photo;
 use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\Auth;
+use App\Models\Comment;
+
 
 
 
@@ -27,7 +29,7 @@ class AdminUsersController extends Controller
         $users = User::all();
 
         return view('admin.users.index',compact('users'));
-
+ 
     }
 
     /**
@@ -72,7 +74,7 @@ class AdminUsersController extends Controller
         }
         User::create($input);
         Session::flash('created_user','User has been created');
-        return redirect('admin/users');
+        return redirect('/admin');
 
     }
 
@@ -110,26 +112,33 @@ class AdminUsersController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(UserEditRequest $request, $id)
+    public function update(Request $request, $id)
     {
         //
-        $user = User::findOrFail($id);
+        // $user = Auth::user();
+        $user = User::findOrFail($request->id);
 
         if(trim($request->password) == ''){
             $input = $request->except('password');
         }else{
             $input = $request->all();
-            $input['password'] = bcrypt($request->password);
+            $input['password'] = bcrypt(trim($request->password));
         }
 
 
 
         if($file = $request->file('photo_id')){
+            if($user->photo_id != 0) {
+                unlink(public_path() . $user->photo->file);
+            }
+            
             $name = time() . $file->getClientOriginalName();
 
             $file->move('images',$name);
 
             $photo = Photo::create(['file'=>$name]);
+
+            Comment::where('email',$user->email)->update(['photo'=>'/images/' . $name]);
 
             $input['photo_id'] = $photo->id;
         }
@@ -137,7 +146,12 @@ class AdminUsersController extends Controller
 
         Session::flash('updated_user','User has been updated');
 
-        return redirect('admin/users');
+        // User::findOrFail($id)->update($request->all());
+
+        return redirect('/admin');
+
+  
+       
     }
 
     /**
@@ -146,18 +160,48 @@ class AdminUsersController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request,$id)
     {
         //
-
         $user = User::findOrFail($id);
-
-        unlink(public_path() . $user->photo->file);
+        if($user->photo_id != 0){
+             unlink(public_path() . $user->photo->file);
+        }
 
         $user->delete();
 
         Session::flash('deleted_user','User has been deleted.');
-        return redirect('/admin/users');
+        // return redirect('/admin');
+
+
+    }
+
+    public function active(Request $request,$id){
+        $user = User::findOrFail($id);
+        $user->update(['is_active'=>$request->is_active]);
+     
+    }
+
+
+
+    public function deleteSelectUser(Request $request){
+        foreach($request->ids as $id){
+            User::where('id',$id)->delete(); 
+        }
+    }
+
+
+    public function activeSelectUser(Request $request){
+        foreach($request->ids as $id){
+            User::where('id',$id)->update(['is_active'=>1]);
+        }
+    }
+
+    public function profile(){
+        $user = Auth::user();
+        return view('admin.users.profile',compact('user'));
     }
 }
+
+
